rename integration test, break up test function
